Update 04.DataRecords.md
Some minor changes, I inserted a "^" in "m2" and exchanged "conversely" with "In cntrast," and some linguistic improvements.
I also added the reference: "We recommend for applications that focus on particular vegetation types to retrieve them from the current sPlot version, following sPlot's Governance and Data Property Rules (www.idiv.de/sPlot)."
In addition, I feel that in the caption of Fig. 2, we miss a reference for the Whittaker biomes, as well as the reference to the R package of Valentin Stefan.