Update 03.Methods.md
A few additional edits for that section that you may deem useful. By the way, I wonder if we could not also provide the code that Tarek wrote, using Rcpp to speed up computation time in C++, so that the reader could also use and play with this code. What do you think? I can send you all the R-scripts that Tarek wrote. Tarek wrote them as standalone functions to be sourced in any independent R script. Thus it could be pretty useful.