Update 03.Methods.md
I addition to my PR this morning, I do have one important comment about this Method section. You mentioned in the text of that section 858 PC1-PC2 grid cells with more than 50 vegetation plots, but how many PC1-PC2 grid cells in total do we have with vegetation plots therein? It might be worth mentionning somewhere the total number of PC1-PC2 grid cells for wich we have vegetation plot data. Also, and related to that, you mentioned later 2,380 PC1-PC2 grid cells (11.7% of total) that had one more confidential vegetation plot. Are you sure about that number of PC1-PC2 grid cells (n = 2,380)? Are you actually refering to the PC1-PC2 grid cells here when mentioning this number 2,380? Isn't it the number of vegetation plots? Sorry, I am a bit confused by this very last sentence you wrote in the sub-section entitled "Permission to release the data as open access". Based on this number 2,380, which is 11.7% of the total, it means a total of 20,341 PC1-PC2 grid cells, is that right? Or does that number refers to the number of vegetation plots? Do we actually need this last sentence in this sub-section entitled "Permission to release the data as open access"?