Update 05.DatabaseOrganization.md
Be careful, you mentioned "ownership" in the header matrix but this variable is not listed in Table 2! Maybe worth adding and listing this variable in Table2, no? Otherwise, simply remove "ownership" from the list of variables provided as exemples for the header matrix. Indeed, this kind of ownership information seems to be provided anyway in the metadata matrix and thus maybe better to only refer to ownership there in the metadata matrix, no? By the way, in Table 2, you could refer to Table 1 for the cells at the crossing between the "Range/Levels" column and the "GIVD_ID" + "Dataset" rows.