Update 06.TechnicalValidation.md
A few edits. Note also that you mention BioTIME in the "Usage Notes" but without refering to the actual paper, maybe worth adding the reference here, no? In the section entitled "Code Availability", we could also mention (and thus provide) the R scipts writtent by Tarek Hattab for the resampling procedure. What do you think? I think it could be of interest to the reader as there are a lot of standalone functions that Tarek built to run the resampling procedure much faster through Rcpp (C++ code in the background).