Minor minor revision and a few comments
I made one or two very minor revision(s) and have the following comments.
"we removed vegetation plots: from wetlands;" -> so no wetland plots are in the db? If yes please make that perfectly clear and I think it would be good to explain the reader why. If no, please make more clear which plots are removed.
"while avoiding peculiar and rare communities, which may represent outliers. " -> but these are often the most valuable for biodiversity analyses? Maybe elaborate tot he reader why this is the smartest choice?