Minor revisions
A few minor revisions and a comment:
If there's ever going to be a version 2.0 one could consider not doing this and just leaving it to the users if the want to do that or not. Analysis og orchids for example might be difficult without subspecies info:
"All taxa originally denoted at taxonomic ranks lower than species were aggregated at species level. "