Update 02.BackgroundSummary.md
I think the reason why this database is created should be put more forward: E.g. "Here, we provide an open-access data set composed of 91,056 vegetation plots, that represent the entire environmental space covered by the sPlot database, to provide a standardized dataset for ecological research." Can be different though, but I think it should be clear to the reader immoderately why this dataset is needed and why its different to sPlot.