Update 01.abstract.md
I have written an alternative abstract, which I think may be an improvement. The original one seems to me to lack key aspects of the rationale, value and description of our contribution. I have added those things in and trimmed out stuff I thought was less important in an abstract, keeping to the 170 word limit. (This is from Richard Field - not sure how clear that is if I don't say so!)