Update 03.Methods.md
In addition to my minor proposed edits, in the section 'Permission to release...' I think we should make it clearer how the open dataset ended up with 91,205 vegetation plots. It is far from obvious from the numbers in this paragraph how we ended up with that number!
In the 'Trait information' section, I find the description of gap-filling confusing.  For any given plot, what trait values are there?  Is it individually measured traits or species-level values?  Not clear, though the passage describing CWM and CWV calculations suggests species-level values.  In that case, what is meant by the sentence "Gap-filling was performed at the level of individual observations"?