Update 04.DataRecords.md
In addition to my edits, I have a comment that some of the numbers are hard to understand.  Specifically:
"sPlot Open contains a relatively balanced number of forest (n = 25,832) vs. non forest (n = 38,203) vegetation plots, with a minor proportion of plots remaining unassigned (n = 10,050)."  These numbers add up to about 74,000 plots, which begs the question of why the other 18,000 plots are neither forest nor non-forest.  I think we need clearer explanation of this, and also of many of the rows in Table 2 that have similar issues.
In Table 2, for me there is not enough information here to understand many of the variables.  Is there a source of more detailed info anywhere?  If so, we should point to it clearly.  Examples of questions arising from Table 2:
* Why are there ~500 plots without a 'continent' assignation?
* Why do 30,000 plots not have a size (Releve_area)?
* Why do 80,000 plots not have either true or false for Herbs_identified?
* Why do 35,000 plots not have either true or false for is_forest?
* Why do 2,000 plots not have either true or false for is_nonforest?
* Why do 180 plots have no entry for Plant_recorded?
And so on.  These things need to be explained somewhere, and it is not clear to me that they are.