Update 03.Methods.md
I am very concerned about applying the gap-filling method to assign traits from TRY to New Zealand species.  Given our high rate of endemism -- 80% of the flora -- what are the implications of using this approach?  It's not as if trait values may even exist for closely related taxa. I found the cited papers for the method quite dense and it will take me a lot of effort to get my head around this method.  But I would be worried if there is some kind of bias in the trait values and then naive users of the dataset will just accept them and come up with wonky predictions and conclusions for our part of the globe.
   I have encountered a different sort of problem when NZ species occurrence data were intersected with a global database of bark thickness.  In this example, they only retained species that were represented in the database and these were all introduced to NZ (northern hemisphere conifers).  This led to the completely wrong conclusion that NZ is dominated by thick-barked trees which would imply fire adaptation, whereas the opposite is true.  I assume the gap-filling technique is meant to overcome this.  I think there needs to be a bit more description provided here about how this gap filling works and what the limitations might be.  Alternatively, it could be useful to present a table of countries and the % of the species that required gap filling to derive traits.