Update 04.DataRecords.md
I am concerned about 
- the immense range of plot sizes, which contain unreasonably small as well as large "plots" - please list proportions of size classes in Table 2.
- inclusion of plots with selective recording of vascular plant species - in any case, list the proportions of plots in Table 2; crucial advantages of vegetation plots mentioned in Background & Summary (e.g. true absences) do not pertain to such selective plots; diversity measures and CWM of traits of such plots are incompatible with complete plots; they also interfere with resampling through distorted dissimilarity patterns etc. I can see they are better then nothing in undersampled biomes, but having them in the database poses the risk of misleading results.
Are Whittaker biomes the best choice for Figure 2, when so many plots (temperate-montane and alpine?) are way outside of their climatic domain? A frequency distribution of plots across biomes would certainly be important (cf. variable Biome in Table 2).
In Braun-Blanquet relevés cover of species is not relative, but always refers to projection on the plot area.