Update 02.AlternativeAbstract.md
Some edits on the abstract. One major important addition I made to the abstract is to specify that sPlotOpen also offers the possibility to select three environmentally balanced datasets of about 50k plots each. I think it is a valuable and important point to mention because we are selling sPlotOpen as an environmentally-balanced dataset (cf. the title). Hence, it has to be explicitly mentionned in the abstract. To me this is even more important that putting forward that sPlotOpen contains almost 100k plots.