Update 30.Methods.md
I think we need to carefully explain in more details in the methods section how many plots were selected by each resampling iteration separately (cf. my suggested edits in the subsection entitled "Permission to release the data as open access"). We also need to carefully check the numbers here annd provide the total number of unique vegetation plost when pooling all three resampling iterations together. Is it 107,238 unique vegetation plots when pooling all three resampling iterations together? I mean, it was not clear to me how you got this number and I assumed this is the unique plot IDs when pooling all thre iterations together, right? About the reserve, it is a bit confusing in the end. I guess the reserve is only useful if we want to provide a "consensus" resampled datasets, right? If we don't provide such a consensus in the end, do we really need to explain and provide this "reserve"? Maybe we just need to explain that the resampling iteration #1 is not the most optimal one to use in the end because a lot of plots are not available as open access (those not granted by data contributors from iteration #1) in the end and thus iteration #2 or #3 could be better in that respect (lower losses). Do you have the exact number of plots not granted in the end per resampling iteration? Maybe this is a useful infoormation to provide here as well, no?