Update 30.Methods.md
I updated the initial sample size of the three resampling iterations by looking at the data and then I used the unique() function in R to check the number of non overlapping plot and indeed I found 107,238, which I now better undrestand where it comes from. About the 3,150 plots from the reserve (coming from resampling iterations #2 and #3, right?), are they included in the display of figures 1 and 2 (n = 49,787)? I guess not because when I do the difference between the initial size of resampling iteration #1 (n = 56,486) and the number of plots not granted as open access (n = 6,625), I find 49,861 plots, which differs slightly from the 49,787 plots in figure 1. Why?