Update 40.DataRecords.md
How do you get the total of 95,104 vegetation plots? I mean, if I do the difference 107,238 minus 12,035 (not granted as open access) I get a total of 95,203. The difference is small but still... Where does that come from? About the resampling iteration #1, does figure 2 show the set of plots without or with the reserve? I assume it is without the reserve, right? Maybe specify clearly in the caption of figure 2 the size of the resampling datatset.